feat(ui): add back to top floating button (Fixes #1415)
