--- components/BackToTop.js
@@ -1,32 +1,0 @@
-"use client";
-
-import { useState, useEffect } from "react";
-import { ChevronUp } from "lucide-react";
-
-export default function BackToTop() {
-  const [isVisible, setIsVisible] = useState(false);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      setIsVisible(window.scrollY > 300);
-    };
-    window.addEventListener("scroll", handleScroll, { passive: true });
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
-
-  const scrollToTop = () => {
-    window.scrollTo({ top: 0, behavior: "smooth" });
-  };
-
-  if (!isVisible) return null;
-
-  return (
-    <button
-      onClick={scrollToTop}
-      aria-label="Back to top"
-      className="fixed bottom-24 right-6 z-50 p-3 rounded-full bg-gradient-to-r from-accent to-purple-600 text-white shadow-lg hover:shadow-accent/40 hover:scale-110 transition-all duration-300 animate-fadeIn"
-    >
-      <ChevronUp className="h-5 w-5" />
-    </button>
-  );
-}
